run OF on just pairwise

+++ scripts/coding_gene_families/prepare.OrthoFinder.files.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-export cluster=$1
+export dataset=$1
+export cluster=$2
 
 ##########################################################################
 
@@ -14,8 +15,9 @@ if [ ${cluster} = "pbil" ]; then
     export pathTools=/beegfs/home/necsulea/Tools/OrthoFinder/tools
 fi
 
+export pathGeMoMa=${path}/results/CDS_annotation
 export pathEnsemblProteins=${path}/data/protein_sequences/Ensembl103
-export pathResults=${path}/results/gene_families/OrthoFinder
+export pathResults=${path}/results/gene_families/OrthoFinder_${dataset}
 export pathScripts=${path}/scripts/coding_gene_families
 
 ##########################################################################
@@ -28,18 +30,45 @@ fi
 
 ##########################################################################
 
-for file in `ls  ${pathEnsemblProteins} `
-do
-    export sp=`echo ${file} | cut -f 1 -d '.'`
-    
-    if [ -e ${pathEnsemblProteins}/primary_transcripts/${file} ]; then
-	echo "primary transcripts already done"
-    else
-	python ${pathTools}/primary_transcript.py ${pathEnsemblProteins}/${file}
-    fi
+if [ ${dataset} = "Ensembl103" ]; then
 
-    ln -s ${pathEnsemblProteins}/primary_transcripts/${file} ${pathResults}/${sp}.fa
-done
+    for file in `ls  ${pathEnsemblProteins} `
+    do
+	export sp=`echo ${file} | cut -f 1 -d '.'`
+	
+	if [ -e ${pathEnsemblProteins}/primary_transcripts/${file} ]; then
+	    echo "primary transcripts already done"
+	else
+	    python ${pathTools}/primary_transcript.py ${pathEnsemblProteins}/${file}
+	fi
+	
+	ln -s ${pathEnsemblProteins}/primary_transcripts/${file} ${pathResults}/${sp}.fa
+    done
+fi
+
+##########################################################################
+
+if [ ${dataset} = "GeMoMa_pairwise" ]; then
+    for sp in Chicken Duck
+    do
+	export pathAnnot=${pathGeMoMa}/${sp}/GeMoMa/combined
+	export file=combined_protein_sequences.faa
+
+	if [ -e ${pathAnnot}/${file} ]; then
+	    if [ -e ${pathAnnot}/primary_transcripts/${file} ]; then
+		echo "primary transcripts already done"
+	    else
+		python ${pathTools}/primary_transcript.py ${pathAnnot}/${file}
+	    fi
+	    
+	    ln -s ${pathEnsemblProteins}/primary_transcripts/${file} ${pathResults}/${sp}.fa
+	    
+	else
+	    echo "cannot find protein sequences! "${pathAnnot}/${file} 
+	    exit
+	fi
+    done
+fi
 
 ##########################################################################
